+++ apps/relayer/src/store/sql/models/user_decrypt_req_model.rs
@@ -15,7 +15,7 @@ pub struct UserDecryptReq {
     pub id: i32,
     pub ext_reference_id: Uuid,
     pub int_indexer_id: Vec<u8>,
-    pub gw_reference_id: Option<i32>,
+    pub gw_reference_id: Option<i64>,
     pub req: Value,
     pub req_status: ReqStatus,
     pub gw_req_tx_hash: Option<String>,
