+++ apps/relayer/src/store/sql/repositories/input_proof_repo.rs
@@ -56,7 +56,7 @@ impl InputProofRepository {
         &self,
         int_request_id: Uuid,
         gw_req_tx_hash: &str,
-        gw_reference_id: i32,
+        gw_reference_id: i64,
     ) -> Result<u64> {
         let result = sqlx::query!(
             r#"
@@ -109,7 +109,7 @@ impl InputProofRepository {
     /// Returns the int_request_id.
     pub async fn accept_and_complete_input_proof_req(
         &self,
-        gw_reference_id: i32,
+        gw_reference_id: i64,
         res: serde_json::Value,
         gw_response_tx_hash: &str,
     ) -> Result<Uuid> {
@@ -139,7 +139,7 @@ impl InputProofRepository {
     /// Returns the int_request_id.
     pub async fn reject_and_complete_input_proof_req(
         &self,
-        gw_reference_id: i32,
+        gw_reference_id: i64,
         res: serde_json::Value,
         gw_response_tx_hash: &str,
     ) -> Result<Uuid> {
